logout after erasure

+++ Model/Customer/Anonymize/Processor/CustomerDataProcessor.php
@@ -11,12 +11,14 @@ use Magento\Customer\Api\CustomerRepositoryInterface;
 use Magento\Customer\Model\CustomerRegistry;
 use Magento\Framework\Api\SearchCriteriaBuilder;
 use Magento\Framework\App\Config\ScopeConfigInterface;
+use Magento\Framework\Exception\InputException;
 use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Exception\NoSuchEntityException;
+use Magento\Framework\Exception\State\InputMismatchException;
 use Magento\Sales\Api\Data\OrderInterface;
+use Magento\Sales\Api\Data\OrderSearchResultInterface;
 use Magento\Sales\Api\OrderRepositoryInterface;
 use Magento\Store\Model\ScopeInterface;
-use Opengento\Gdpr\Model\Customer\Anonymize\AccountBlocker;
 use Opengento\Gdpr\Service\Anonymize\AnonymizerInterface;
 use Opengento\Gdpr\Service\Erase\ProcessorInterface;
 
@@ -28,11 +30,6 @@ final class CustomerDataProcessor implements ProcessorInterface
      * @var AnonymizerInterface
      */
     private $anonymizer;
-
-    /**
-     * @var AccountBlocker
-     */
-    private $accountBlocker;
 
     /**
      * @var CustomerRepositoryInterface
@@ -61,7 +58,6 @@ final class CustomerDataProcessor implements ProcessorInterface
 
     public function __construct(
         AnonymizerInterface $anonymizer,
-        AccountBlocker $accountBlocker,
         CustomerRepositoryInterface $customerRepository,
         OrderRepositoryInterface $orderRepository,
         SearchCriteriaBuilder $criteriaBuilder,
@@ -69,7 +65,6 @@ final class CustomerDataProcessor implements ProcessorInterface
         ScopeConfigInterface $scopeConfig
     ) {
         $this->anonymizer = $anonymizer;
-        $this->accountBlocker = $accountBlocker;
         $this->customerRepository = $customerRepository;
         $this->orderRepository = $orderRepository;
         $this->criteriaBuilder = $criteriaBuilder;
@@ -84,28 +79,14 @@ final class CustomerDataProcessor implements ProcessorInterface
     public function execute(int $customerId): bool
     {
         $isRemoved = false;
+
         try {
-            if ($this->shouldRemoveCustomerWithoutOrders()) {
-                $this->criteriaBuilder->addFilter(OrderInterface::CUSTOMER_ID, $customerId);
-                $orderList = $this->orderRepository->getList($this->criteriaBuilder->create());
-
-                if (!$orderList->getTotalCount()) {
-                    $isRemoved = $this->customerRepository->deleteById($customerId);
-                }
+            if ($this->shouldRemoveCustomerWithoutOrders() && !$this->fetchOrdersList($customerId)->getTotalCount()) {
+                $isRemoved = $this->customerRepository->deleteById($customerId);
             }
-
-            // Make sure, we don't work with cached customer data, because
-            // saving cached customers may "de-anonymize" related data
-            // like addresses
-            $this->customerRegistry->remove($customerId);
-
             if (!$isRemoved) {
-                $this->accountBlocker->invalid($customerId);
-                $this->customerRepository->save(
-                    $this->anonymizer->anonymize($this->customerRepository->getById($customerId))
-                );
+                $this->anonymizeCustomer($customerId);
             }
-
         } catch (NoSuchEntityException $e) {
             return false;
         }
@@ -113,11 +94,30 @@ final class CustomerDataProcessor implements ProcessorInterface
         return true;
     }
 
+    private function fetchOrdersList(int $customerId): OrderSearchResultInterface
+    {
+        $this->criteriaBuilder->addFilter(OrderInterface::CUSTOMER_ID, $customerId);
+
+        return $this->orderRepository->getList($this->criteriaBuilder->create());
+    }
+
+    /**
+     * @param int $customerId
+     * @throws LocalizedException
+     * @throws NoSuchEntityException
+     * @throws InputException
+     * @throws InputMismatchException
+     */
+    private function anonymizeCustomer(int $customerId): void
+    {
+        $this->customerRegistry->remove($customerId);
+        $this->customerRepository->save(
+            $this->anonymizer->anonymize($this->customerRepository->getById($customerId))
+        );
+    }
+
     private function shouldRemoveCustomerWithoutOrders(): bool
     {
-        return $this->scopeConfig->isSetFlag(
-            self::CONFIG_PATH_ERASURE_REMOVE_CUSTOMER,
-            ScopeInterface::SCOPE_STORE
-        );
+        return $this->scopeConfig->isSetFlag(self::CONFIG_PATH_ERASURE_REMOVE_CUSTOMER, ScopeInterface::SCOPE_STORE);
     }
 }
